fastapi-postgres-app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
pytest
httpx
asyncpg
email-validator